# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Asset Management System API (ISO 27001)
  version: 1.0.0
  description: RESTful API for IT Asset & ISMS Management with ISO 27001 compliance
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api/v1
    description: Development
paths:
  /phase6/resources:
    get:
      summary: List available Phase 6 resources
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Resource keys for generic Phase 6 API }
  /phase6/{resource}:
    get:
      summary: List Phase 6 records with filters
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: resource, in: path, required: true, schema: { type: string } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: overdue, in: query, required: false, schema: { type: boolean } }
      responses:
        '200': { description: Paginated Phase 6 records }
    post:
      summary: Create Phase 6 record
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: resource, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '201': { description: Created record }
  /phase6/{resource}/{id}:
    get:
      summary: Get Phase 6 record
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: resource, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Phase 6 record }
    patch:
      summary: Update Phase 6 record
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: resource, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated record }
  /phase6/{resource}/export:
    get:
      summary: Export Phase 6 records as JSON or CSV
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: resource, in: path, required: true, schema: { type: string } }
        - { name: format, in: query, required: false, schema: { type: string, enum: [json, csv] } }
      responses:
        '200': { description: Persisted ExportJob }
  /phase6/corrective-actions/from-source:
    post:
      summary: Create CAPA from Audit, Incident, Risk, Control or Supplier
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Corrective action }
  /phase6/workflows/start:
    post:
      summary: Start a generic workflow instance
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Workflow instance }
  /phase6/workflows/{id}/transition:
    post:
      summary: Transition a generic workflow instance
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated workflow instance }
  /phase6/reports/run:
    post:
      summary: Run and persist a Phase 6 report
      tags: [Phase6]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: ReportRun }
...
  /risks/risk-controls:
    post:
      summary: Link risk to a concrete control implementation
      tags: [Risks]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskControlInput'
      responses:
        '201': { description: RiskControl link created }
  /risks/risk-control-assessments:
    post:
      summary: Assess a RiskControl for a specific RiskAssessmentVersion
      tags: [Risks]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskControlAssessmentInput'
      responses:
        '201': { description: Versioned risk-control assessment created }
  /risks/assessment-versions/{id}/close:
    post:
      summary: Close an immutable risk assessment version
      tags: [Risks]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: RiskAssessmentVersion closed }
  /controls/tests:
    post:
      summary: Create a control implementation test with generic evidence links
      tags: [Controls]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlTestInput'
      responses:
        '201': { description: Control test created }
...
  /risks:
    get:
      summary: List risks (paginated, filterable)
      tags: [Risks]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  # ==================== Controls ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RiskControlInput:
      type: object
      required: [riskId, controlImplementationId, role, mitigationDimension]
      description: Canonical Risk-Control link; replaces deprecated direct arrays such as existingControls, relatedRiskIds, controlIds and riskIds.
      properties:
        riskId: { type: string, format: uuid }
        controlImplementationId: { type: string, format: uuid }
        role: { type: string, enum: [preventive, detective, corrective, recovery, compensating] }
        mitigationDimension: { type: string, enum: [likelihood, impact, both] }
        isKeyControl: { type: boolean, default: false }
        status: { type: string, default: active }
    RiskControlAssessmentInput:
      type: object
      required: [riskControlId, riskAssessmentVersionId, effectivenessStatus, justification, assessedBy]
      properties:
        riskControlId: { type: string, format: uuid }
        riskAssessmentVersionId: { type: string, format: uuid }
        effectivenessStatus: { type: string, enum: [not_verified, ineffective, partially_effective, effective] }
        effectivenessRating: { type: integer, minimum: 0, maximum: 100 }
        likelihoodReduction: { type: integer }
        impactReduction: { type: integer }
        justification: { type: string }
        assessedBy: { type: string }
        evidenceLinks:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceLinkInput'
    ControlTestInput:
      type: object
      required: [controlImplementationId, testType, testedBy, result]
      properties:
        controlImplementationId: { type: string, format: uuid }
        testType: { type: string }
        testMethod: { type: string }
        testedBy: { type: string }
        testedAt: { type: string, format: date-time }
        result: { type: string }
        effectivenessRating: { type: integer, minimum: 0, maximum: 100 }
        findings: { type: string }
        evidenceRequired: { type: boolean }
        nextTestDate: { type: string, format: date-time }
        evidenceLinks:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceLinkInput'
    EvidenceLinkInput:
      type: object
      required: [evidenceId]
      properties:
        evidenceId: { type: string, format: uuid }
        relationType: { type: string, default: supports }
...
    PaginatedResponse:
      type: object
      properties:
        data: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }

# ==================== Phase 8: API, Betrieb und Nachweise ====================

  # ==================== Health & System (Phase 8) ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== Phase 8 Schemas ====================
...